+++ src/features/course-view/data/index.js
@@ -1,0 +1,2 @@
+export { reducer } from './slice';
+export * from './selectors';